refactor: move unexposed files into internal
