+++ v-next/hardhat-ignition/src/internal/tasks/deploy.ts
@@ -21,12 +21,12 @@ import debug from "debug";
 import Prompt from "prompts";
 
 import { HardhatArtifactResolver } from "../../hardhat-artifact-resolver.js";
-import { PrettyEventHandler } from "../../ui/pretty-event-handler.js";
-import { bigintReviver } from "../../utils/bigintReviver.js";
-import { loadModule } from "../../utils/load-module.js";
-import { readDeploymentParameters } from "../../utils/read-deployment-parameters.js";
-import { resolveDeploymentId } from "../../utils/resolve-deployment-id.js";
-import { shouldBeHardhatPluginError } from "../../utils/shouldBeHardhatPluginError.js";
+import { PrettyEventHandler } from "../ui/pretty-event-handler.js";
+import { bigintReviver } from "../utils/bigintReviver.js";
+import { loadModule } from "../utils/load-module.js";
+import { readDeploymentParameters } from "../utils/read-deployment-parameters.js";
+import { resolveDeploymentId } from "../utils/resolve-deployment-id.js";
+import { shouldBeHardhatPluginError } from "../utils/shouldBeHardhatPluginError.js";
 
 interface TaskDeployArguments {
   modulePath: string;
