+++ v-next/hardhat-ignition/src/internal/ui/helpers/calculate-batch-display.ts
