+++ v-next/hardhat-ignition/src/internal/ui/helpers/calculate-deploying-module-panel.ts
