+++ v-next/hardhat-ignition/src/internal/ui/helpers/calculate-deployment-complete-display.ts
