+++ v-next/hardhat-ignition/src/internal/ui/helpers/calculate-deployment-status-display.ts
