+++ v-next/hardhat-ignition/src/internal/ui/helpers/calculate-list-transactions-display.ts
