+++ v-next/hardhat-ignition/src/internal/ui/helpers/calculate-starting-message.ts
