+++ v-next/hardhat-ignition/src/internal/ui/helpers/was-anything-executed.ts
