+++ v-next/hardhat-ignition/src/internal/ui/pretty-event-handler.ts
