+++ v-next/hardhat-ignition/test/ui/helpers/calculate-batch-display.ts
@@ -1,12 +1,12 @@
-import type { UiFuture, UiState } from "../../../src/ui/types.js";
+import type { UiFuture, UiState } from "../../../src/internal/ui/types.js";
 
 import { assert } from "chai";
 
-import { calculateBatchDisplay } from "../../../src/ui/helpers/calculate-batch-display.js";
+import { calculateBatchDisplay } from "../../../src/internal/ui/helpers/calculate-batch-display.js";
 import {
   UiFutureStatusType,
   UiStateDeploymentStatus,
-} from "../../../src/ui/types.js";
+} from "../../../src/internal/ui/types.js";
 
 import { testFormat } from "./test-format.js";
